requirements.txt
customtkinter>=5.2.0
PyAudioWPatch>=0.2.12.8
websockets>=12.0
python-dotenv>=1.0.0
Pillow>=10.0.0
